Added support for authn. (#2)
* Added support for authn.

This is done by using go-containerregistry's kubenernetes authn module.

To enable the pod to access the secret, the following constructs are used:
* a service account for the pod
* a role that allows for accessing secrets
* a role binding that bind the role to the service account

Note that the secret is not part of this repo, any user is expected
to first create the imge pull secret as we can't easily provision that
via helm.

* updated the readme

* spelling errors

* added linter and build step. Added tons of comments

* go get -u

* go get -u

* errors in workflow files

* correct step in workflow

* minor style fix

* minor cleanups

* updated error message

* spelling errors